+++ wsidicom/file/__init__.py
@@ -14,6 +14,5 @@
 
 """Module for handling DICOM WSI instances read from file."""
 
-from wsidicom.file.wsidicom_file_base import OffsetTableType
 from wsidicom.file.wsidicom_file_source import WsiDicomFileSource
-from wsidicom.file.wsidicom_file_writer import WsiDicomFileWriter
+from wsidicom.file.wsidicom_file_target import WsiDicomFileTarget