+++ wsidicom/source.py
@@ -18,7 +18,7 @@ from typing import Iterable
 from wsidicom.graphical_annotations import AnnotationInstance
 from wsidicom.instance import WsiDataset, WsiInstance
 
-"""A Source provdes image instances that can be opened by WsiDicom. Implement this
+"""A Source provides image instances that can be opened by WsiDicom. Implement this
 class to extend support to other DICOM sources or to read other WSI formats.
 """