+++ wsidicom/web/__init__.py
@@ -12,7 +12,7 @@
 #    See the License for the specific language governing permissions and
 #    limitations under the License.
 
-"""Module for handling DICOM WSI instances read using DICOM Web."""
+"""Module for handling DICOM WSI instances read using DICOMWeb."""
 
 from wsidicom.web.wsidicom_web_client import WsiDicomWebClient
 from wsidicom.web.wsidicom_web_source import WsiDicomWebSource